+++ examples/simple_audio_volume_example.py
@@ -2,13 +2,13 @@
 Per session GetMute() SetMute() using ISimpleAudioVolume.
 """
 
-from pycaw.pycaw import AudioUtilities, ISimpleAudioVolume
+from pycaw.pycaw import AudioUtilities
 
 
 def main():
     sessions = AudioUtilities.GetAllSessions()
     for session in sessions:
-        volume = session._ctl.QueryInterface(ISimpleAudioVolume)
+        volume = session.SimpleAudioVolume
         print("volume.GetMute(): %s" % volume.GetMute())
         volume.SetMute(1, None)