+++ pycaw/api/audioclient/depend.py
@@ -1,13 +1,13 @@
 from ctypes import Structure
-from ctypes.wintypes import WORD
+from ctypes.wintypes import DWORD, WORD
 
 
 class WAVEFORMATEX(Structure):
     _fields_ = [
         ("wFormatTag", WORD),
         ("nChannels", WORD),
-        ("nSamplesPerSec", WORD),
-        ("nAvgBytesPerSec", WORD),
+        ("nSamplesPerSec", DWORD),
+        ("nAvgBytesPerSec", DWORD),
         ("nBlockAlign", WORD),
         ("wBitsPerSample", WORD),
         ("cbSize", WORD),