+++ tests/test_examples.py
@@ -20,11 +20,12 @@ class TestExamples:
             audio_endpoint_volume_example.main()
         output = out.getvalue()
         lines = output.split("\n")
-        assert lines[0] == "volume.GetMute(): 0"
-        assert lines[1] == "volume.GetMasterVolumeLevel(): -20.0"
-        assert lines[2] == "volume.GetVolumeRange(): (-95.25, 0.0, 0.75)"
-        assert lines[3] == "volume.SetMasterVolumeLevel()"
-        assert lines[4] == "volume.GetMasterVolumeLevel(): -20.0"
+        assert lines[0].startswith('Device found: ')
+        assert lines[1] == "volume.GetMute(): 0"
+        assert lines[2] == "volume.GetMasterVolumeLevel(): -20.0"
+        assert lines[3] == "volume.GetVolumeRange(): (-95.25, 0.0, 0.75)"
+        assert lines[4] == "volume.SetMasterVolumeLevel()"
+        assert lines[5] == "volume.GetMasterVolumeLevel(): -20.0"
 
     def test_simple_audio_volume_example(self):
         with captured_output() as (out, err):